Split components in files
